Add python venv support, handle signal and shutdown on cancel ai operations.
If you cancel an detection/reid, we don't shutdown cleanly; the python process
stays running. If we add a signal handler that doesn't help as the signal
can't arrive on detection when we're doing mp.pool.map() as that blocks the
main thread and the signal handler can't run. So make that map_async. Even
so, when using Conda it appears that map_async blocks! So setup a python
venv and use standard python; the signal arrives then and we shutdown cleanly.

I still left Conda workable; we detect if we're in a conda env, and will use
that if present, so that researchers can still use Conda if they like. We
should use standard Python for production though, as Conda had bad signal
handling with multiprocessing map.