Rework Python detect/reid code to use pyinstaller binaries.
Make a dedicated cpu and gpu pyinstaller binary which is started in a child
process by the main process rather than running from inside the venv. This
pyinstaller runtime has to be built and copied into the resources and
referenced in the JS code so that it's bundled by electron-builder when we
make packages.

I am able to create a Mac DMG with this which is stand alone and works as
expected. The only issue is the stock photos no longer load over web in
electron in preview/prod mode due to a CSP issue. I had resolved that in
dev mode, not sure why it's not working in prod.